+++ R/radsl.R
@@ -87,12 +87,20 @@ radsl <- function(N = 400, # nolint
     ) %>%
       sample_fct(N, prob = c(.55, .23, .16, .05, .004, .003, .002, .002)),
     TRTSDTM = as.POSIXct(
-      sys_dtm + sample(seq(0, study_duration_secs), size = N, replace = TRUE),
+      sys_dtm + sample(seq(0, study_duration_secs / 8), size = N, replace = TRUE),
       origin = "1970-01-01"
     ),
+    TRT01SDTM = TRTSDTM,
+    AP01SDTM = TRT01SDTM,
     RANDDT = as.Date(.data$TRTSDTM) - floor(stats::runif(N, min = 0, max = 5)),
-    st_posixn = as.numeric(.data$TRTSDTM),
-    TRTEDTM = as.POSIXct(.data$st_posixn + study_duration_secs, origin = "1970-01-01"),
+    st_posixn = as.numeric(.data$TRT01SDTM),
+    TRT01EDTM = as.POSIXct(.data$st_posixn + study_duration_secs / 4, origin = "1970-01-01"),
+    AP01EDTM = TRT01EDTM,
+    st_posixn_2 = as.numeric(.data$TRT01EDTM),
+    TRT02SDTM = as.POSIXct(.data$st_posixn_2 + study_duration_secs / 8, origin = "1970-01-01"),
+    AP02SDTM = TRT02SDTM,
+    st_posixn_3 = as.numeric(.data$TRT02SDTM),
+    TRTEDTM = as.POSIXct(.data$st_posixn_3 + study_duration_secs / 2, origin = "1970-01-01"),
     STRATA1 = c("A", "B", "C") %>% sample_fct(N),
     STRATA2 = c("S1", "S2") %>% sample_fct(N),
     BMRKR1 = stats::rchisq(N, 6),
@@ -109,6 +117,8 @@ radsl <- function(N = 400, # nolint
     dplyr::mutate(ACTARMCD = .data$ARMCD) %>%
     dplyr::mutate(TRT01P = .data$ARM) %>%
     dplyr::mutate(TRT01A = .data$ACTARM) %>%
+    dplyr::mutate(TRT02P = sample(.data$ARM)) %>%
+    dplyr::mutate(TRT02A = sample(.data$ACTARM)) %>%
     dplyr::mutate(ITTFL = factor("Y")) %>%
     dplyr::mutate(SAFFL = factor("Y")) %>%
     dplyr::arrange(.data$st_posixn)
@@ -131,8 +141,11 @@ radsl <- function(N = 400, # nolint
       st_posixn >= quantile(st_posixn)[2] & st_posixn <= quantile(st_posixn)[3] ~ as.POSIXct(NA, origin = "1970-01-01"),
       TRUE ~ TRTEDTM
     )) %>%
-    dplyr::mutate(TRTEDTM = as.POSIXct(.data$TRTEDTM, origin = "1970-01-01")) %>%
-    dplyr::select(-.data$TRTEDTM_discon)
+    dplyr::mutate(TRTEDTM = as.POSIXct(.data$TRTEDTM, origin = "1970-01-01"), # nolint
+                  AP02EDTM = TRTEDTM,
+                  TRT02EDTM = TRTEDTM
+                  ) %>% # nolint
+    dplyr::select(-.data$TRTEDTM_discon) # nolint
 
   ADSL <- ADSL %>% # nolint
     dplyr::mutate(EOSDT = as.Date(.data$TRTEDTM)) %>%
@@ -207,7 +220,7 @@ radsl <- function(N = 400, # nolint
       DCSREAS == "DEATH" ~ DTHDT,
       TRUE ~ as.Date(.data$TRTEDTM) + floor(stats::runif(N, min = 10, max = 30))
     )) %>%
-    dplyr::select(-.data$st_posixn)
+    dplyr::select(-.data$st_posixn, -.data$st_posixn_2, -.data$st_posixn_3)
 
   # add random ETHNIC (Ethnicity)
   ADSL <- ADSL %>% # nolint