+++ NEWS.md
@@ -23,6 +23,9 @@
 * Add variables `CMSTRTPT`, `CMENRTPT`, `ADURN`, and `ADURU` to `adcm`.
 * Add variables `AESCONG`, `AESDISAB`, `AESHOSP`, `AESLIFE`, and `AESMIE` to `adae`.
 * Add variable `ADTHAUT` to `adsl`.
+* Add variables `TRT02A`, `TRT02P`, `TRT02SDTM`, `TRT02EDTM`, `AP02SDTM`, `AP02EDTM`, `TRT01SDTM`, `TRT01EDTM`, 
+  `AP01SDTM`,and `AP01EDTM` to `adsl`.
+* Enhanced variables `TRTSDTM` as `TRT01SDTM`, `TRTEDTM` as `TRT02EDTM`.
 
 # random.cdisc.data 0.3.13